This commit was generated by cvs2svn to compensate for changes in r5823, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@5824 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/prdevice.c
@@ -39,6 +39,7 @@
 #include "print.h"
 #include "resources.h"
 #include "serial.h"
+#include "types.h"
 #include "utils.h"
 #include "vdrive.h"
 
@@ -173,7 +174,7 @@ static int write_pr(void *var, BYTE byte, int secondary)
 	if (er < 0) return er;
     }
 
-    return print_putc(currfd, byte);
+    return print_putc(currfd, (BYTE)byte);
 }
 
 static int close_pr(void *var, int secondary)
